backend/requirements.txt
flask
flask-jwt-extended
flask-cors
supabase
python-dotenv
scikit-learn
werkzeug
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers